Separate API tests into their own modules
