+++ tests/test_wesim.py
@@ -1,15 +1,4 @@
 import pandas as pd
-import pytest
-
-from datahub import data as dt
-
-
-@pytest.fixture
-def wesim_input_data():
-    """The filepath for the test version of the wesim data."""
-    from datahub.wesim import read_wesim
-
-    return read_wesim("tests/data/wesim_example.xlsx")
 
 
 def test_read_wesim(wesim_input_data):
@@ -84,11 +73,3 @@ def test_get_wesim(mocker, wesim_input_data):
     assert pd.DataFrame(**wesim["Regions"]).shape == (30, 10)
     assert pd.DataFrame(**wesim["Interconnector Capacity"]).shape == (4, 2)
     assert pd.DataFrame(**wesim["Interconnectors"]).shape == (25, 3)
-
-
-def test_get_wesim_api(client, mocker, wesim_input_data):
-    """Test get_wesim returns a dictionary with appropriate DataFrames."""
-    with mocker.patch("datahub.wesim.read_wesim", return_value=wesim_input_data):
-        response = client.get("/wesim")
-
-    assert response.json()["data"] == dt.wesim_data
